refactor(openomni): merge resident runtime satellites into runtime.ts
runtime-abort/runtime-agent-config/runtime-types importers were all
cluster-internal; runtime-types held mutable activation state only
runtime.ts touches. Doc-comment path references updated.
